backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pytesseract==0.3.10
pdf2image==1.17.0
PyPDF2==3.0.1
Pillow==9.5.0
python-dotenv==1.0.1
Werkzeug==3.0.1